textService/requirements.txt
fastapi
uvicorn
sqlalchemy.orm
.database
requests
openai
dotenv
google.generativeai
